Corrected a error in the way the panel was being positioned.
Created : debug, log, message and die routine to help debug
and started replacing raw printfs to the them.

Improved the way the border radius is changed when it is to big.

Depracated center contraint for vertical positioning.

Fixed some bad ifs in the panel.c `init_panel_size_and_position()'.